Move chat types to a different file
