+++ WalletWasabi/BuyAnything/BuyAnythingManager.cs
@@ -20,49 +20,7 @@ namespace WalletWasabi.BuyAnything;
 // Event that is raised when we detect an update in the server
 public record ConversationUpdateEvent(Conversation Conversation, DateTimeOffset LastUpdate);
 
-public abstract record DataCarrier
-{
-	public static DataCarrier NoData = new NoData();
-}
-
-public record Invoice(string Bip21Link, decimal Amount, string BitcoinAddress) : DataCarrier;
-
-public record OfferItem(float Quantity, string Description, float UnitPrice, float TotalPrice);
-public record OfferCarrier(IEnumerable<OfferItem> Items) : DataCarrier;
-
-public record TrackingCodes(IEnumerable<string> Codes) : DataCarrier;
-public record AttachmentLinks(IEnumerable<string> Codes) : DataCarrier;
-public record NoData : DataCarrier;
-
-public record ChatMessage(bool IsMyMessage, string Message, bool IsUnread, ChatMessageMetaData MetaData);
-
-public record SystemChatMessage(string Message, DataCarrier Data, bool IsUnread, ChatMessageMetaData MetaData)
-	: ChatMessage(false, Message, IsUnread, MetaData);
-
 public record Country(string Id, string Name);
-
-public record ChatMessageMetaData(ChatMessageMetaData.ChatMessageTag Tag)
-{
-	public static readonly ChatMessageMetaData Empty = new(ChatMessageTag.None);
-
-	public enum ChatMessageTag
-	{
-		None = 0,
-
-		AssistantType = 11,
-		Country = 12,
-
-		FirstName = 21,
-		LastName = 22,
-		StreetName = 23,
-		HouseNumber = 24,
-		PostalCode = 25,
-		City = 26,
-		State = 27,
-
-		PaymentInfo = 31,
-	}
-}
 
 [Flags]
 internal enum ServerEvent
